+++ CHANGES.rst
@@ -8,6 +8,7 @@ Changelog
 - Add classifiers for Plone 5.
   [agitator]
 - Add python3 support.
+- Handle case when image has no wrapper tag.
   [cekk]
 
 1.0.1 (2018-10-19)